Remove contact_profiles, make Person+Channel the sole contact system
Drop all contact_profiles and contact_profile_exceptions code:
- Remove table definitions from pg_schema, models from pg_models
- Remove 10 repository methods from user_repository
- Remove old CRUD handlers, rename file to person_handlers
- Remove legacy fallbacks from bridge, tool_call_utils, context
- Remove old routes from main.rs, add /api/persons/search route
- Update elevenlabs, scheduler, dashboard, profile_handlers
